+++ src/spring/HISTORY.md
@@ -1,5 +1,15 @@
 Release History
 ===============
+1.7.0
+---
+* Print application logs when create/update deployment
+* Bypass jar check for enterprise tier
+* Add java runtime check for jar file
+
+1.6.8
+---
+* Add detail description to Default for argument `--backend-protocol`.
+
 1.6.7
 ---
 * Change all Azure Spring Apps API version to 2022-11-01-preview.